requirements.txt
accelerate==0.32.1
datasets==2.20.0
mosaicml-cli==0.5.34
mosaicml-streaming==0.8.1
ninja==1.11.1.1
numpy==1.26.3
packaging==24.1
pandas==2.2.1
protobuf==4.25.3
python-dateutil==2.9.0
regex==2023.12.25
sentencepiece==0.1.99
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.4
transformers==4.44.2
triton==3.0.0
wandb==0.17.3
zstandard==0.22.0
zstd==1.5.5.1


